momentum_bot_railway/requirements.txt
websocket-client==1.6.1
requests==2.31.0
